+++ careerconnect/accounts/apps.py
@@ -3,7 +3,3 @@ from django.apps import AppConfig
 class AccountsConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'accounts'
-
-    def ready(self):
-        import accounts.signals
-
