+++ careerconnect/accounts/templates/accounts/profile.html
@@ -67,7 +67,7 @@
             </p>
             <div class="button-row">
                 <a href="/" class="btn-cancel">Cancel</a>
-                <button type="submit" class="btn-save">Save Changes</button>
+                <button type="submit" class="btn-save" style="height: 44px">Save Changes</button>
             </div>
         </form>
     </div>
