+++ careerconnect/careerconnect/static/css/style.css
@@ -49,7 +49,7 @@ p {
 }
 
 a:visited {
-    color:#00cfd5
+    color:#ffffff
 }
 
 nav {
