+++ careerconnect/jobs/templates/jobs/job_form.html
@@ -49,7 +49,7 @@
 
             <div class="button-row">
                 <a href="{% url 'jobs:recruiter_jobs' %}" class="btn-cancel">Cancel</a>
-                <button type="submit" class="btn-save">Save Job</button>
+                <button type="submit" class="btn-save" style="height: 44px">Save Job</button>
             </div>
         </form>
     </div>
